% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCompoundDbPackage.R
\name{createCompoundDb}
\alias{createCompoundDb}
\title{Create a CompoundDb database}
\usage{
createCompoundDb(x, metadata, path = ".")
}
\arguments{
\item{x}{\code{data.frame} or \code{tbl} with the compound annotations. See
description for details.}

\item{metadata}{\code{data.frame} with metadata information. See description for
details.}

\item{path}{\code{character(1)} with the path to the directory where the database
file should be written. Defaults to the current directory.}
}
\value{
A \code{character} with the database name (invisibly).
}
\description{
\code{createCompoundDb} creates a \code{SQLite}-based \code{CompoundDb} object/database
from a compound resource provided as a \code{data.frame} or \code{tbl}.
An additional \code{data.frame} providing metadata information is mandatory.
Required columns for the \code{data.frame} providing the compound information are:

The metadata \code{data.frame} is supposed to have two columns named \code{"name"} and
\code{"value"} providing the following minimal information as key-value pairs:
\itemize{
\item \code{"source"}: the source from which the data was retrieved (e.g. \code{"HMDB"}).
\item \code{"url"}: the url from which the original data was retrieved.
\item \code{"source_version"}: the version from the original data source
(e.g. \code{"v4"}).
\item \code{"source_date"}: the date when the original data source was generated.
\item \code{"organism"}: the organism. Should be in the form \code{"Hsapiens"} or
\code{"Mmusculus"}.
}
}
\details{
Metadata information is also used to create the file name for the database
file. The name starts with \code{"CompoundDb"}, followed by the organism, the
data source and its version. A compound database file for HMDB version 4
with human metabolites will thus be named: \code{"CompoundDb.Hsapiens.HMDB.v4"}.
}
\examples{

## Read compounds for a HMDB subset
fl <- system.file("extdata/hmdb/hmdb_sub.xml", package = "PeakABro")
cmps <- generate_hmdb_tbl(fl)

## Create a metadata data.frame for the compounds.
metad <- data.frame(name = c("source", "url", "source_version",
    "source_date", "organism"), value = c("HMDB", "http://www.hmdb.ca",
    "v4", "2017-08-27", "Hsapiens"))

## Create a SQLite database in the temporary folder
db_f <- createCompoundDb(cmps, metadata = metad, path = tempdir())

## connect to the database and query it's tables using RSQlite
library(RSQLite)
con <- dbConnect(dbDriver("SQLite"), db_f)

dbGetQuery(con, "select * from metadata")
dbGetQuery(con, "select * from compound")
}
\author{
Johannes Rainer
}
