% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cmp_tbl.R
\name{generate_hmdb_tbl}
\alias{generate_hmdb_tbl}
\title{Generate a table with HMDB compounds}
\usage{
generate_hmdb_tbl(file)
}
\arguments{
\item{file}{\code{character} with the name(s) of xml files downloaded from HMDB,
or the file name of a HMD file in SDF format.
For xml files: can be a single xml file (containing all compounds from
an HMDB release) or the names of several xml files, each supposed to
provide data for one compound.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with general compound information (one row per
compound):
\itemize{
\item \code{id}: the HMDB ID of the compound.
\item \code{name}: the compound's name.
\item \code{inchi}: the inchi of the compound.
\item \code{formula}: the chemical formula.
\item \code{mass}: the compound's mass. The value from HMDB's
\code{monisotopic_molecular_weight} is returned.
}
}
\description{
\code{generate_hmdb_tbl} processes one or more HMDB xml files
(downloaded from http://http://www.hmdb.ca/) or an HMDB file in SDF
format and extracts general compound information.
}
\note{
At present only a subset of the available data provided by HMDB are
extracted. Also, compounds for which no mass is provided are not
reported.
}
\examples{

## Extract compound information from a single xml file from HMDB.
fl <- system.file("extdata/hmdb/HMDB0000001.xml", package = "PeakABro")
hmdb <- generate_hmdb_tbl(fl)
colnames(hmdb)
hmdb

## Extract compound information from a (subset of a) HMDB release xml file.
fl <- system.file("extdata/hmdb/hmdb_sub.xml", package = "PeakABro")
hmdb <- generate_hmdb_tbl(fl)
colnames(hmdb)
hmdb

## Extract compound information from a HMDB file in SDF format
fl <- system.file("extdata/hmdb/hmdb_sub.sdf", package = "PeakABro")
hmdb <- generate_hmdb_tbl(fl)
hmdb

## Extract the masses
hmdb$mass
}
\seealso{
Other compound table creation functions: \code{\link{generate_lipidblast_tbl}},
  \code{\link{generate_lipidmaps_tbl}}
}
\author{
Johannes Rainer
}
