% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoundDb.R
\name{CompoundDb}
\alias{CompoundDb}
\title{Simple compound (metabolite) databases}
\description{
\code{CompoundDb} objects provide access to general (metabolite)
compound annotations along with \emph{metadata} information such as the
annotation's source, date and release version. The data is stored
internally in a database (usually an SQLite database).
}
\author{
Johannes Rainer
}
