% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompoundDb.R, R/CompoundDb-methods.R
\docType{methods}
\name{CompoundDb}
\alias{CompoundDb}
\alias{CompoundDb-class}
\alias{show,CompoundDb-method}
\alias{dbconn,CompoundDb-method}
\alias{show}
\alias{dbconn}
\alias{CompoundDb}
\alias{dbconn,CompoundDb-method}
\title{Simple compound (metabolite) databases}
\usage{
dbconn(x)
show(object)

CompoundDb(x)

\S4method{dbconn}{CompoundDb}(x)
}
\arguments{
\item{x}{For \code{CompoundDb}: \code{character(1)} with the file name of the SQLite
compound database. Alternatively it is possible to provide the connection
to the database with parameter \code{x}.\preformatted{For all other methods: a `CompoundDb` object.
}}

\item{object}{For all methods: a \code{CompoundDb} object.}
}
\description{
\code{CompoundDb} objects provide access to general (metabolite) compound
annotations along with \emph{metadata} information such as the annotation's
source, date and release version. The data is stored internally in a
database (usually an SQLite database).

`CompoundDb` *constructs* a `CompoundDb` object by connecting
    to the provided database file.

`dbconn` returns the connection (`DBIConnection`) to the
    database.
}
\details{
\code{CompoundDb} objects should be created using the constructor function
\code{CompoundDb} providing the name of the (SQLite) database file providing
the compound annotation data.
}
\examples{

## Create a small CompoundDb from a provided HMDB subset
cmps <- generate_hmdb_tbl(system.file("extdata/hmdb/hmdb_sub.xml",
    package = "PeakABro"))
metad <- data.frame(name = c("source", "url", "source_version",
    "source_date", "organism"),
    value = c("sub_HMDB", "http://www.hmdb.ca", "4", "2017", "Hsapiens"),
    stringsAsFactors = FALSE)
## Create the SQLite database:
db_file <- createCompoundDb(cmps, metadata = metad, path = tempdir())

## Create a CompoundDb object
cmp_db <- CompoundDb(db_file)
cmp_db
}
\seealso{
\code{\link[=createCompoundDb]{createCompoundDb()}} for the function to create a SQLite compound
database.
}
\author{
Johannes Rainer
}
